Improve documentation, support prefix removal from basenames
The documentation describing the way in which Tenji objects differ from
Jekyll::Page was incomplete and lacked clarity. This commit improves the
explanations.

In preparing the documentation, it was realised that Tenji did not support
the removal of prefixes (specifically ordinal patterns) from basenames. A user
of Tenji may wish to include an ordinal pattern at the beginnig of a file's
basename in order to control the ordering. A similar mechanism is available for
gallery names. However, unlike gallery names, Tenji preserved the ordinal
pattern in the output files. This was inconsistent with how it worked for
directory names and could lead to confusion. Changes were made to remove these
prefixes on output.