Add ThumbFile documentation, change ThumbFile initialiser
In the course of documenting the ThumbFile class, it became apparent
that there was no good reason for why the ThumbFile initialiser allowed
for a reference to the source image to be included. The other
Processable objects all anticipate cross references being added after
instantiation. Indeed, cover images cannot be instantiated with a
sensible reference since the source may not be known by Tenji at the
time of instantiation. That parameter was removed and the calls to
ThumbFile::new were updated accordingly.

An errant `puts` debugging statement was also removed.