Switch YARD to Commonmarker, add Templating documentation
By default, YARD uses the redcarpet Markdown parsing engine. For various
reasons, CommonMark provides a preferable engine and this commit adds
a plugin to provide this support.

In addition, this commit adds documentation about templating. Tenji
requires the user to provide templates for the pages Tenji creates and
the documentation provides information to enable a user to write the
necessary templates.

Finally, changes were made to the Configuring documentation to take
into account the fact that CommonMark does not natively support
definition lists (which had been used previously).